Refactor Storable in its own file. Add documentation
