+++ Pantry/Pantry.swift
@@ -8,89 +8,6 @@
 
 import Foundation
 
-// MARK: the protocol your stuff should adhere to
-
-public protocol Storable {
-    init(warehouse: JSONWarehouse)
-    func toDictionary() -> [String: AnyObject]
-}
-
-extension Storable {
-    // Adapted from @IanKeen's https://gist.github.com/IanKeen/3a6c3b9a42aaf9fea982
-    func toDictionary() -> [String: AnyObject] {
-        let mirror = Mirror(reflecting: self)
-        return mirror.children.reduce([:]) { result, child in
-            guard let key = child.label else { return result }
-            
-            let childMirror = Mirror(reflecting: child.value)
-            if let style = childMirror.displayStyle where style == .Collection {
-                // collections need to be unwrapped, children tested and
-                // toDictionary called on each
-                let converted: [AnyObject] = childMirror.children
-                    .filter { $0.value is Storable || $0.value is AnyObject }
-                    .map { collectionChild in
-                        if let convertable = collectionChild.value as? Storable {
-                            return convertable.toDictionary()
-                        } else {
-                            return collectionChild.value as! AnyObject
-                        }
-                    }
-                return combine(result, addition: [key: converted])
-                
-            } else {
-                // non-collection types, toDictionary or just cast default types
-                // optionals need to be checked and unwrapped
-                let childMirror = Mirror(reflecting: child.value)
-
-                if let value = child.value as? Storable {
-                    return combine(result, addition: [key: value.toDictionary()])
-                } else if let value = child.value as? AnyObject {
-                    return combine(result, addition: [key: value])
-                } else if childMirror.displayStyle == .Optional {
-                    // yes, this is how you detect and unwrap an Optional
-                    // disguised as an Any
-                    if childMirror.children.count != 0 {
-                        let (_, some) = childMirror.children.first!
-                        if let some = some as? Storable {
-                            return combine(result, addition: [key: some.toDictionary()])
-                        }
-                    }
-                } else {
-                    // throw an error? not a type we support
-                }
-            }
-            
-            return result
-        }
-    }
-    
-    // convenience for combining dictionaries
-    func combine(from: [String: AnyObject], addition: [String: AnyObject]) -> [String: AnyObject] {
-        var result = [String: AnyObject]()
-        [from, addition].forEach { dict in
-            dict.forEach { result[$0.0] = $0.1 }
-        }
-        return result
-    }
-}
-
-public enum StorageExpiry {
-    case Never
-    case Seconds(NSTimeInterval)
-    case Date(NSDate)
-    
-    func toDate() -> NSDate {
-        switch self {
-        case Never:
-            return NSDate.distantFuture()
-        case Seconds(let timeInterval):
-            return NSDate(timeIntervalSinceNow: timeInterval)
-        case Date(let date):
-            return date
-        }
-    }
-}
-
 // MARK: the main public class
 
 public class Pantry {
